Fix initial values for inherited properties.
It looks like values are set correctly to start with, but corrupted in `GenerateScriptResource_SCI0` after calling `context.FixupSinksAndSources()`.

I noticed that `CSCOObjectProperty` has two constructors. One does have an initializer list, which means the instance fields are initialized with arbitrary values. This meant that `_fNeedsReloc` would sometimes be initialized as a non-zero value, which would always require object properties to be relocated, which seems to cause the corrupt property values.

Adding a member initializer seems to avoid "relocating" the property values, and I can see that inherited values are now correct when disassembling a script.